Rework the Request System and Centralize it

+++ src/main/java/fr/iglee42/notenoughchests/utils/ModAbbreviation.java
@@ -22,44 +22,41 @@ public class ModAbbreviation {
 
     public static final File BACKUP_FILE = new File(FMLPaths.CONFIGDIR.get().resolve("nec").toFile(), "abbreviations_backup.json");
 
-    public static boolean API_ONLINE = true;
-
     public static void init(){
-        try {
-            URL url = URI.create("https://iglee.fr:3000/modAbbreviations").toURL();
-            HttpURLConnection con = (HttpURLConnection)url.openConnection();
-            con.setRequestMethod("GET");
-            con.setConnectTimeout(5000);
-            con.setReadTimeout(1000);
-            con.setInstanceFollowRedirects(false);
-            int status = con.getResponseCode();
-            if (status == 200) {
-                BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
-                StringBuilder json = new StringBuilder();
-                String inputLine;
-                while ((inputLine = in.readLine()) != null) {
-                    try {
-                        json.append(inputLine);
-                    } catch (Exception ignored) {
+        if (RequestsUtils.API_ONLINE) {
+            try {
+                HttpURLConnection con = RequestsUtils.sendRequest("modAbbreviations", "GET", 0);
+                int status = con.getResponseCode();
+                if (status == 200) {
+                    BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
+                    StringBuilder json = new StringBuilder();
+                    String inputLine;
+                    while ((inputLine = in.readLine()) != null) {
+                        try {
+                            json.append(inputLine);
+                        } catch (Exception ignored) {
+                        }
                     }
-                }
-                JsonObject obj = new Gson().fromJson(json.toString(), JsonObject.class);
-                obj.entrySet().forEach(e -> {
-                    abbreviations.put(e.getKey(), e.getValue().getAsString());
-                });
+                    JsonObject obj = new Gson().fromJson(json.toString(), JsonObject.class);
+                    obj.entrySet().forEach(e -> {
+                        abbreviations.put(e.getKey(), e.getValue().getAsString());
+                    });
 
-                new File(BACKUP_FILE.getParent()).mkdirs();
-                if (BACKUP_FILE.exists()) BACKUP_FILE.delete();
-                try (FileWriter fw = new FileWriter(BACKUP_FILE)) {
-                    fw.write(new GsonBuilder().setPrettyPrinting().create().toJson(obj));
-                } catch (IOException ignored) {
-                }
+                    new File(BACKUP_FILE.getParent()).mkdirs();
+                    if (BACKUP_FILE.exists()) BACKUP_FILE.delete();
+                    try (FileWriter fw = new FileWriter(BACKUP_FILE)) {
+                        fw.write(new GsonBuilder().setPrettyPrinting().create().toJson(obj));
+                    } catch (IOException ignored) {
+                    }
 
-                in.close();
-                con.disconnect();
+                    in.close();
+                    con.disconnect();
+                }
+            } catch (Exception ignored) {
+                NotEnoughChests.LOGGER.error("Error while asking the abbreviations,some chests may not work correctly !");
+                loadBackup();
             }
-        } catch (Exception ignored) {
-            API_ONLINE = false;
+        } else {
             NotEnoughChests.LOGGER.error("API isn't online, some chests may not work correctly !");
             loadBackup();
         }
@@ -80,7 +77,7 @@ public class ModAbbreviation {
                 obj.entrySet().forEach(e->{
                     abbreviations.put(e.getKey(),e.getValue().getAsString());
                 });
-                NotEnoughChests.LOGGER.info("Abbreviations loaded with backup file ! ");
+                NotEnoughChests.LOGGER.info("Abbreviations loaded from backup file ! ");
                 return;
             } catch (Exception ignored){}
         }
